+++ Conv/Item/Item.swift
@@ -8,36 +8,6 @@
 
 import UIKit
 
-public protocol ItemDelegatable: Reusable {
-    func configureCell(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath)
-    func sizeFor(collectionView: UICollectionView, indexPath: IndexPath) -> CGSize?
-    
-    func canMoveItem(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    
-    func willDisplay(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath)
-    func didEndDisplay(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath)
-    
-    func shouldHighlight(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    func didHighlight(collectionView: UICollectionView, indexPath: IndexPath)
-    func didUnhighlight(collectionView: UICollectionView, indexPath: IndexPath)
-    
-    func shouldSelect(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    func shouldDeselect(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    func didSelect(collectionView: UICollectionView, indexPath: IndexPath)
-    func didDeselect(collectionView: UICollectionView, indexPath: IndexPath)
-    
-    func shouldShowMenu(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    func canPerformAction(collectionView: UICollectionView, action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) -> Bool?
-    func performAction(collectionView: UICollectionView, action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?)
-    
-    func canFocusItem(collectionView: UICollectionView, indexPath: IndexPath) -> Bool?
-    
-    func shouldSpringLoadItem(collectionView: UICollectionView, indexPath: IndexPath, context: UISpringLoadedInteractionContext) -> Bool?
-    
-    func targetIndexPathForMoveFromItem(collectionView: UICollectionView, originalIndexPath: IndexPath, proposedIndexPath: IndexPath) -> IndexPath?
-}
-
-
 public class Item<Cell: UICollectionViewCell>: Reusable {
     public typealias ItemArgument = (item: Item<Cell>, collectionView: UICollectionView, indexPath: IndexPath)
     public typealias PerformActionArgument = (item: Item<Cell>, collectionView: UICollectionView, action: Selector, indexPath: IndexPath, sender: Any?)
@@ -136,79 +106,5 @@ extension Item {
     
     public func targetIndexPathForMoveFromItem(_ closure: @escaping ((Item<Cell>, UICollectionView, IndexPath, IndexPath) -> IndexPath)) {
         self.targetIndexPathForMoveFromItem = closure
-    }
-}
-
-extension Item: ItemDelegatable {
-    public func configureCell(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath) {
-        configureCell?(cell as! Cell, (self, collectionView, indexPath))
-    }
-    
-    public func sizeFor(collectionView: UICollectionView, indexPath: IndexPath) -> CGSize? {
-        return sizeFor?((self, collectionView, indexPath)) ?? size
-    }
-    
-    public func canMoveItem(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return canMoveItem?((self, collectionView, indexPath))
-    }
-    
-    public func willDisplay(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath) {
-        willDisplay?(cell as! Cell, (self, collectionView, indexPath))
-    }
-    
-    public func didEndDisplay(collectionView: UICollectionView, cell: UICollectionViewCell, indexPath: IndexPath) {
-        didEndDisplay?(cell as! Cell, (self, collectionView, indexPath))
-    }
-    
-    public func shouldHighlight(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return shouldHighlight?((self, collectionView, indexPath))
-    }
-    
-    public func didHighlight(collectionView: UICollectionView, indexPath: IndexPath) {
-        didHighlight?((self, collectionView, indexPath))
-    }
-    
-    public func didUnhighlight(collectionView: UICollectionView, indexPath: IndexPath) {
-        didUnhighlight?((self, collectionView, indexPath))
-    }
-    
-    public func shouldSelect(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return shouldSelect?((self, collectionView, indexPath))
-    }
-    
-    public func shouldDeselect(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return shouldDeselect?((self, collectionView, indexPath))
-    }
-    
-    public func didSelect(collectionView: UICollectionView, indexPath: IndexPath) {
-        didSelect?((self, collectionView, indexPath))
-    }
-    
-    public func didDeselect(collectionView: UICollectionView, indexPath: IndexPath) {
-        didDeselect?((self, collectionView, indexPath))
-    }
-    
-    public func shouldShowMenu(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return shouldShowMenu?((self, collectionView, indexPath))
-    }
-    
-    public func canPerformAction(collectionView: UICollectionView, action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) -> Bool? {
-        return canPerformAction?((self, collectionView, action, indexPath, sender))
-    }
-    
-    public func performAction(collectionView: UICollectionView, action: Selector, forItemAt indexPath: IndexPath, withSender sender: Any?) {
-        performAction?((self, collectionView, action, indexPath, sender))
-    }
-    
-    public func canFocusItem(collectionView: UICollectionView, indexPath: IndexPath) -> Bool? {
-        return canFocusItem?(self, collectionView, indexPath)
-    }
-    
-    public func shouldSpringLoadItem(collectionView: UICollectionView, indexPath: IndexPath, context: UISpringLoadedInteractionContext) -> Bool? {
-        return shouldSpringLoadItem?(self, collectionView, indexPath, context)
-    }
-    
-    public func targetIndexPathForMoveFromItem(collectionView: UICollectionView, originalIndexPath: IndexPath, proposedIndexPath: IndexPath) -> IndexPath? {
-        return targetIndexPathForMoveFromItem?(self, collectionView, originalIndexPath, proposedIndexPath)
     }
 }
